Initial support for bitmap based fbconsd
To use, comment all but fbconsd in the embedbins for the "Graphics stuff"
section and ensure FREETYPE_CONSOLE is set to "no" in the GNUMakefile.  Using
the full "Graphics stuff" section with FREETYPE_CONSOLE set to "yes" should
give the original style freetype console.  Also, make sure vesa is enabled in
csetup.c.  The startup time is significantly improved with the non-freetype
version.