Checkpoint work in progress.
The libhtc_ril.so appears to finish sending AT commands to the modem
via the smd interface and eventually hangs on a condvar. My only guess
right now is that the qmi interfaces need to be available for whatever
event it is waiting on. The RIL says that the modem is still at state 1,
i.e. is unavailable/hasn't finished initialization.